Moved RestoreUserAsync and RestoreUsersAsync to Command.
